rework BBMASK to use regex
we use BBMASK to blacklist all packages within the ADI upstream layer.
This is done, as they are mostly outdated. However, we do need the
u-boot and kernel related recipes.

This commit changes the BBMASK from individual paths to a regex. This is
to make sure, that new, unwanted recipes do not pop-up when upstream is
updated.