Don't assume build username within the docker container
Previously we assumed the build user is named "builder" within the KAS
container. However, this could be open to change. As the user is added
at runtime, we can mount the SSH folder to /etc/skel/.ssh instead, which
is the skeleton profile for newly created users. By doing so, we do not
rely on the naming of the build user