Modify environment variables in docker-compose
Previously one had to specify the current user ID when using
docker-compose. Now, docker-compose will default to UID=1000 and
GID=1000, if variables are not provided.

Additionally, the path to a folder containing an id_rsa file for cloning from
private repositories can now be specified using the SSH_DIR variable